-just clear code add generic view instead usual view
