+++ courierservice/courierservice/settings.py
@@ -103,4 +103,3 @@ EMAIL_FILE_PATH = os.path.join(BASE_DIR, 'sent_emails')
 LOGIN_URL = 'users:login'
 LOGIN_REDIRECT_URL = 'orders:index'
 
-# CSRF_FAILURE_VIEW = 'core.views.csrf_failure'
