+++ courierservice/templates/orders/product_list.html
@@ -1,38 +1,54 @@
 {% extends "base.html" %}
 {% block title %}
-    Список продуктов
+    Главная страница
 {% endblock %}
 {% block header %}
 {% endblock %}
 {% block content %}
-    <div id="sidebar">
-        <h3>Categories</h3>
-        <ul>
-            <li {% if not category %}class="selected"{% endif %}>
-                <a href="{% url "orders:product_list:" %}">All</a>
-            </li>
-            {% for c in categories %}
-                <li {% if category.slug == c.slug %}class="selected"{% endif %}>
-                    <a href="{{ c.get_absolute_url }}">{{ c.name }}</a>
-                </li>
-            {% endfor %}
-        </ul>
-    </div>
-    <div class="card-group">
-        <div class="row">
-            {% for product in products %}
-                <div class="col-md-4">
-                    <div class="card" style="max-width: 500px; height: 550px;">
-                        <img class="card-img-top" src="{{ product.image.url }}" alt="Card image cap" width="200"
-                             height="300">
-                        <div class="card-body">
-                            <h5 class="card-title">{{ product.name }}</h5>
-                            <p class="card-text">{{ product.description }}</p>
-                            <a href="#" class="btn btn-primary">Go somewhere</a>
+        ОООО Это главная страница
+        <a href="{% url 'orders:new_order' %}">Сделать заказ</a>
+        <div class="container">
+            <div class="row">
+                {% for product in products %}
+                    <div class="col-lg-4 .col-sm-6">
+                        <div class="card" style="max-width: 500px; height: 550px;">
+                            <img class="card-img-top" src="{{ product.image.url}}" alt="Card image cap" width="200" height="300">
+                            <div class="card-body">
+                                <h5 class="card-title">{{ product.name }}</h5>
+                                <p class="card-text">{{ product.description }}</p>
+                                <a href="#" class="btn btn-primary">Go somewhere</a>
+                            </div>
                         </div>
                     </div>
-                </div>
-            {% endfor %}
+                {% endfor %}
+            </div>
         </div>
-    </div>
 {% endblock %}
+<link rel="stylesheet" type="text/css" href="product.css">
+    <section class="main-content">
+        <div class="container">
+            <div class="row">
+                {% for product in products %}
+                    <div class="col-lg-4 .col-sm-6">
+                    <div class="product-card">
+                        <div class="product-thumb">
+                            <a href="#"><img src="{{ product.image.url}}" alt=""></a>
+                        </div>
+                        <div class="product-details">
+                            <h4><a href="#">{{ product.name }}</a></h4>
+                            <p>{{ product.description }}</p>
+                        </div>
+                        <div class="product-bottom-details d-flex justify-content-between">
+                            <div class="product-price">
+                                {{ product.price }}
+                            </div>
+                            <div class="product-links">
+                                <a href="#">Добавить в корзину</a>
+                            </div>
+                        </div>
+                    </div>
+                {% endfor %}
+                </div>
+            </div>
+        </div>
+    </section>
